core: move Message into state_transition.go
